Data.Hexagon

Description

Documentation

data Hexagon ori

The fundamental coordinate type for a 2D hexagonal grid.

Grid orientations

data FlatTop

The orientation in which hexagons have a horizontal edge on the top.

Pictorially:      __
               __/  \__
              /  \__/  \
              \__/  \__/
              /  \__/  \
              \__/  \__/
                 \__/

Constructors

FlatTop

data PointyTop

The orientation in which hexagons have vertical edges on the sides.

Pictorially:
               /\   /\
              /  \ /  \
             |    |    |
             |    |    |
              \  / \  /
               \/   \/

Constructors

PointyTop

Relative directions

data TopBottomDir

Relative directions with respect to "Top-Bottom"-ness.

On a FlatTop, these correspond to the edges:

            Top
            ____
   TopLeft /    \ TopRight
          /      \
          \      /
BottomLeft \____/ BottomRight
           Bottom

On a PointyTop, these correspond to the corners:

           Top
            /\
  TopLeft  /  \ TopRight
          |    |
          |    |
BottomLeft \  / BottomRight
            \/
          Bottom

Constructors

BottomRight

Bottom

BottomLeft

TopLeft

Top

TopRight

data EastWestDir

Relative directions with respect to "East-West"-ness.

On a FlatTop, these correspond to the corners:

NorthWest ____ NorthEast
         /    \
   West /      \  East
        \      /
         \____/
 SouthWest    SouthEast

On a PointyTop, these correspond to the edges:

NorthWest /\ NorthEast
         /  \
        |    |
   West |    | East
         \  /
SouthWest \/ SouthEast

Constructors

East

SouthEast

SouthWest

West

NorthWest

NorthEast

class Orientation ori where

Locks down which relative directions correspond to edges/corners on a given orientation.

Associated Types

type EdgeDir ori :: *

Relative directions of edges on a given hexagon.

type CornerDir ori :: *

Relative directions of corners on a given hexagon.

Methods

neighbor :: EdgeDir ori -> Hexagon ori -> Hexagon ori

Compute the hexagon which borders the argument hexagon on the relative edge.

neighbors :: Hexagon ori -> [Hexagon ori]

Retrieves all neighbors for a hexagon in no particular order.

Inter-hexagon operations

binOp :: (Integer -> Integer -> Integer) -> Hexagon ori -> Hexagon ori -> Hexagon ori

Kind of like a monomorphic zipWith on two Hexagons. In pseudo-Haskell:

let ax = toHexagon Axial
    h  = (x, y ) :: (Integer, Integer)
    h' = (x',y') :: (Integer, Integer)
    f            :: Integer -> Integer -> Integer
binOp f (ax h) (ax h') <-> ax $ (f x x', f y y')

plus :: Hexagon ori -> Hexagon ori -> Hexagon ori

Adds the corresponding components of two Hexagons.

plus = binOp (+)

minus :: Hexagon ori -> Hexagon ori -> Hexagon ori

Subtracts the corresponding components of two Hexagons.

minus = binOp (-)

times :: Hexagon ori -> Hexagon ori -> Hexagon ori

Multiplies the corresponding components of two Hexagons.

times = binOp (*)

Adjacency

dist :: Hexagon ori -> Hexagon ori -> Integer

Computes the distance between two Hexagons. This distance can be thought of the length of the straight-line path between two Hexagons.

Building collections

lineTo :: Hexagon ori -> Hexagon ori -> [Hexagon ori]

Returns a list of all Hexagons on the straight-line path from the first Hexagon to the second Hexagon, including both as endpoints.

range :: Hexagon ori -> Integer -> [Hexagon ori]

All Hexagons within a given distance from the argument Hexagon.

rangeWithBlocked :: Hexagon ori -> Integer -> (Hexagon ori -> Bool) -> [Hexagon ori]

Same as range except it factors in which Hexagons are blocked.

The predicate you pass to this function asks if a given Hexagon is blocked. For example:

let origin                     = toHexagon Axial (0,0)
    blockCircleWithRadius4 hex = dist origin hex < 4
    take10Steps                = 10
    startHex                   = toHexagon Axial (2,2)
in rangeWithBlocked startHex take10Steps blockCircleWithRadius4

In the above example, the function starts at the Hexagon located at Axial(2,2). It takes as many as ten steps from that start location. It skips any Hexagon that is closer than four units from the origin. The result it returns is all of the Hexagons that could have been stepped on.

ringAround :: Hexagon ori -> Integer -> [Hexagon ori]

Retrieves all Hexagons contained in the ring of given radius about the argument Hexagon.

In pseudo-Haskell:

ringAround h i = [h' | h' <- allHexagons, dist h h' == i]

Rotations

data RotateDir

Simple sum type for Clockwise/Counter-clockwise.

Constructors

CW

CCW

rotateAboutOrigin :: RotateDir -> Hexagon ori -> Hexagon ori

Rotate a Hexagon 60 degrees in a given direction about the origin.

rotateAroundPoint :: Hexagon ori -> RotateDir -> Hexagon ori -> Hexagon ori

Rotate a Hexagon 60 degrees in a given direction about a center Hexagon.

let rotationCenter :: Hexagon ori
    pointToRotate  :: Hexagon ori
in rotateAroundPoint rotationCenter CCW pointToRotate

Alternative coordinate systems

class HexCoordTuple t where

A class to structure conversions between coordinate systems

Methods

toHexagon :: t -> (Integer, Integer) -> Hexagon ori

fromHexagon :: t -> Hexagon ori -> (Integer, Integer)

data OddR

The Odd-R Offset coordinate system. See http://www.redblobgames.com/grids/hexagons/

Constructors

OddR

data EvenR

The Even-R Offset coordinate system. See http://www.redblobgames.com/grids/hexagons/

Constructors

EvenR

data OddQ

The Odd-Q Offset coordinate system. See http://www.redblobgames.com/grids/hexagons/

Constructors

OddQ

data EvenQ

The Even-Q Offset coordinate system. See http://www.redblobgames.com/grids/hexagons/

Constructors

EvenQ

data Axial

The Axial Offset coordinate system. See http://www.redblobgames.com/grids/hexagons/

Constructors

Axial